Move search trigger components out of Layout and Home
