+++ src/components/CommandBar/SearchIcon.tsx
